requirements.txt
praw
feedparser
pyyaml
ruamel.yaml